backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
python-dotenv
python-multipart
twilio
gspread 
google-auth